+++ src/Multitenant/ITenantDistributor.cs
@@ -22,7 +22,7 @@ namespace Stashbox.Multitenant
         /// </summary>
         /// <param name="tenantId">The identifier of the tenant.</param>
         /// <param name="tenantConfig">The service configuration of the tenant.</param>
-        void ConfigureTenant(object tenantId, Func<IStashboxContainer, IDisposable> tenantConfig);
+        void ConfigureTenant(object tenantId, Action<IStashboxContainer> tenantConfig);
 
         /// <summary>
         /// Gets a pre-configured <see cref="IDependencyResolver"/> from the distributor which represents a tenant identified by the given id.
@@ -31,10 +31,5 @@ namespace Stashbox.Multitenant
         /// <param name="tenantId">The identifier of the tenant.</param>
         /// <returns>The pre-configured tenant container if it's exist, otherwise null.</returns>
         IDependencyResolver? GetTenant(object tenantId);
-
-        /// <summary>
-        /// Validates the root container and the configured tenants.
-        /// </summary>
-        void Validate();
     }
 }