+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
- 
+
+## [v5.7.0] - 
+### Changed
+- `TenantDistributor` now implements `IStashboxContainer` for easier integration.
+
 ## [v5.6.0] - 2022-12-06
 ### Added
 - `WhenResolutionPathHas()` & `WhenInResolutionPathOf()` registration options for handling more conditional resolution cases. They extend the original *parent type* and *attribute* conditions with inheritance.
@@ -283,6 +287,7 @@ The validation was executed only at the expression tree building phase, so an al
 - Removed the legacy container extension functionality.
 - Removed the support of PCL v259.
 
+[v5.7.0]: https://github.com/z4kn4fein/stashbox/compare/5.6.0...5.7.0
 [v5.6.0]: https://github.com/z4kn4fein/stashbox/compare/5.5.3...5.6.0
 [v5.5.3]: https://github.com/z4kn4fein/stashbox/compare/5.5.2...5.5.3
 [v5.5.2]: https://github.com/z4kn4fein/stashbox/compare/5.5.1...5.5.2